refactor: consolidate history item logic into utility functions
- Moved history item grouping and ownership logic into utility functions for better code organization and reusability.
- Removed redundant implementations from PublicProfile and UserProfile components, streamlining their code.
- Improved maintainability by centralizing history item handling in a single location.
